ls methods confused file names with auth token
The "ls" methods do a quick check to see if the string passed in is an
auth token.
E.g.

testuser.3666.3666.$1$PKCFmMzy$OPpZg/ThBmZe/V8LVPvpi/%

Originally it did so by counting the number of dots (.) in
the string. So a file name such as:

0000m091.1.mod.fits

would also be considered a token and the "ls" command returned with a:
"Not Found"

I added the "is_auth_token" method, which does a thorough regex for what
we expect an auth token to be, making sure that especial cases, such as
anonymous and dldemo tokens are also matched,
and replace the dot counting in the "ls" methods for the "is_auth_token"
check.